fix(scalars): move the hook to better localition in the code
